+++ src/sll/api/vm.c
@@ -5,9 +5,10 @@
 #include <sll/common.h>
 #include <sll/location.h>
 #include <sll/object.h>
-#include <sll/thread.h>
+#include <sll/scheduler.h>
 #include <sll/static_object.h>
 #include <sll/string.h>
+#include <sll/thread.h>
 #include <sll/types.h>
 #include <sll/vm.h>