+++ src/sll/include/sll/_internal/scheduler.h
@@ -6,10 +6,6 @@
 
 
 
-#define SCHEDULER_MAX_THREADS 16384
-
-
-
 typedef __SLL_U32 queue_length_t;