+++ src/sll/include/sll/thread.h
@@ -23,17 +23,6 @@
 
 
 /**
- * \flags var
- * \name sll_current_thread_index
- * \group thread
- * \desc Docs!
- * \type sll_thread_index_t
- */
-__SLL_EXTERNAL extern sll_thread_index_t sll_current_thread_index;
-
-
-
-/**
  * \flags check_output func
  * \name sll_thread_create
  * \group thread