+++ src/sll/lock.c
@@ -3,6 +3,8 @@
 #include <sll/_internal/thread.h>
 #include <sll/common.h>
 #include <sll/memory.h>
+#include <sll/scheduler.h>
+#include <sll/semaphore.h>
 #include <sll/thread.h>
 #include <sll/types.h>