+++ src/sll/vm.c
@@ -23,8 +23,9 @@
 #include <sll/operator.h>
 #include <sll/platform/memory.h>
 #include <sll/sandbox.h>
-#include <sll/thread.h>
+#include <sll/scheduler.h>
 #include <sll/static_object.h>
+#include <sll/thread.h>
 #include <sll/types.h>
 #include <sll/vm.h>